Split ContentView to separate views
