+++ NotesApp/Views/StartScreenView.swift
